refactor: move the connectors list component
