+++ src/routes/index.tsx
@@ -1,6 +1,7 @@
 import { Suspense, lazy } from 'react';
+import { Navigate } from 'react-router-dom';
 const AboutUs = lazy(() => import('@/views/AboutUs'));
-const Dashboard = lazy(() => import('@/views/Dashboard'));
+// const Dashboard = lazy(() => import("@/views/Dashboard"));
 const SignIn = lazy(() => import('@/views/Authentication/SignIn'));
 const SignUp = lazy(() => import('@/views/Authentication/SignUp'));
 const AccountVerify = lazy(() => import('@/views/AccountVerify'));
@@ -30,7 +31,12 @@ export const MAIN_PAGE_ROUTES: MAIN_PAGE_ROUTES_ITEM[] = [
   {
     name: 'Homepage',
     url: '/',
-    component: <SuspenseWithLoader redirectRoute='/'>{<Dashboard />}</SuspenseWithLoader>,
+    component: (
+      <SuspenseWithLoader redirectRoute='/setup/sources'>
+        {/* <Dashboard /> */}
+        <Navigate to='/setup/sources' />
+      </SuspenseWithLoader>
+    ),
   },
   {
     name: 'About Us',
